+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [UNRELEASED]
+* [#524] Added `noreferrer` to Banner links
+
 ## v0.41.x-14
 * [#488] Updated missing proposer address in proposals.
 * [#486] Fixed delegators voting power display on gov proposals