+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [UNRELEASED]
 * [#524] Added `noreferrer` to Banner links
+* [#447] Fixed txs scrolling issue that was causing page to reload and jump to the top when user scrolled to the bottom of the page. 
 
 ## v0.41.x-14
 * [#488] Updated missing proposer address in proposals.